% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_time_series.R
\name{boxplot_time_series}
\alias{boxplot_time_series}
\title{Gulf of Alaska Time Series Boxplot}
\usage{
boxplot_time_series(hist_data, core_stations, plot_type, min_depth = 0,
  max_depth = 100, fastcat_data = FALSE)
}
\arguments{
\item{hist_data}{Supply the path and .csv file name of the historical data
in quotations. The historical data must be in the format created by
the FastrCAT::make_dataframe_fc, which is the same format EcoDAAT exports.
Historic data must be queried from EcoDAAT and saved as a .csv file. In the
future there will be a direct link to the Oracle database where EcoDAAT data
is housed.}

\item{core_stations}{There are three core areas for the Gulf of Alaska which
have been regularly sampled and are representative of the Gulf of Alaska.
Core areas are all in bottom depth at or greater than 100 and less than 150
meters. Salintiy or temperature data from the surface to 100m will be shown.
core_stations are as follows: "Line 8 FOX", "Line 8", "Semidi Spring" and
"Semidi Summer".
Line 8 FOX is a set of six stations  with a bounding box of \code{57.52, -155.2, 57.72, -154.85}. Line 8 are the 4 core stations of Line 8 FOX. Semidi are
6- 8 stations centrally located in the Semidi area and have been consistently
sampled, with a bounding box of '55.1, -158.5, 55.9, -158.0'.}

\item{plot_type}{will accept on of two quoted characters "temperature" or
"salinity".}

\item{min_depth}{Minimum depth which is shown on the plot. The default is set
as 0 meters.}

\item{max_depth}{Maximum depth which is displayed on the plot. The default
is set at 100 meters.}

\item{fastcat_data}{An optional argument if you want to add the current years
fastcat data. Supply the path and .csv file name to the current years
fastcat data. This must be in the format created by the FastrCAT function
make_dataframe_fc.}
}
\value{
A boxplot of temperature or salinity for each year. The width of the boxplot
are proportional to year sample size. The line inside the boxplot is the median and
the ends of the boxplot corresponds to the 25th and 75th percentiles. The
whiskers extend to the largest or smallest observation greater than or equal
1.5 * Interquartile range. Outliers are circles in orange. The historical median
for all years is red and the standard deviation is a blue dashed line.
be written to the folder designated by the historical data file path. The
plot will be in .png format.
}
\description{
Creates either temperature or salinity plots of core EcoFOCI
stations in the Gulf of Alaska. Each plot displays the average temperature
or salinity for each meter of depth of the core stations for each year for
the months when peak sampling of these regions occured. Line 8 and Semidi
area are most commonly sampled in May and June, which is considered
Spring. Summer sampling in the Gulf of Alaska has been less frequent and
starts in the early 2000's. This summer sampling is in the Semidi core
area, summer is considered August and September.Post 2010, these core
stations were only sampled in odd numbered years. In the future more core
areas will be added.
}
