% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gak1.R
\name{compare_gak1}
\alias{compare_gak1}
\title{Compare Temperature and Salinity to GAK1 line}
\usage{
compare_gak1(current_path, foci_header = TRUE)
}
\description{
Compare Temperature and Salinity to GAK1 line
}
