% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastrcat_doc.R
\docType{package}
\name{FastrCAT}
\alias{FastrCAT}
\alias{FastrCAT-package}
\title{FastrCAT: A package for the Analysis of Oceanographic data from FastCATs.}
\description{
This package makes it easy to use temperature and salinity data collected
by a SeaBrid FastCAT. The package provides functions to read in .up files
and bind them into a single dataframe, provide QA/QC feedback, plot depth
by temperature/salinity for each cast, and maps of these... indicies.
}
