% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataframe_fc.R
\name{make_dataframe_fc}
\alias{make_dataframe_fc}
\title{Create a Dataframe from .up files}
\usage{
make_dataframe_fc(current_path, GE = FALSE, Cruise_report = TRUE)
}
\arguments{
\item{current_path}{The path to directory where all .up files
are located for a cruise.}

\item{GE}{A logical value, Returns the dataframe to the global R environment.
By defulaut it is set to FALSE. Set to TRUE if you would like the data
available in the global environment.}

\item{Cruise_report}{A logical value set to TRUE. When TRUE a cruise report
will be generated. When set to false a cruise report will not be
generated.}
}
\value{
.csv file of all .up file data. An .html file with a cruise
summary and data QAQC output.
}
\description{
This function writes a single data frame in .csv format to file
containing oceanographic data collected by the FastCat during a cruise.
The format and column naming conventions are specific to the needs of EcoDAAT.
This is the primary function of the FastrCAT package and must be run prior to
all other functions. All other functions depend on the data frame generated.
Other outputs of the function are two text files. This first is a cruise
summary.The summary contains basic information about the data and some summary
statistics.The second is a warnings file, which tells users if data or
information is missing incase reprocessing is necessary.
}
