% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fc.R
\name{map_fc}
\alias{map_fc}
\title{Mapping of FastCAT data}
\usage{
map_fc(current_path)
}
\arguments{
\item{current_path}{The path to the directory where the .csv file
generated by make_dataframe_fc is located.}
}
\value{
A station map
}
\description{
Once make_dataframe_fc() has been run, then map_fc can be
used. This function makes a station map from the dataframe generated.
The rendered map will be located in the plot folder within the current
folder.
}
