% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Time Series Plots}
\usage{
plot_time_series(hist_data, core_stations, plot_type,
  fastcat_data = FALSE)
}
\arguments{
\item{hist_data}{Supply the path and .csv file name of the historical data
in quotations. The historical data must be in the format created by
the FastrCAT function make_dataframe_fc.}

\item{core_stations}{There are three core areas for the Gulf of Alaska which
have been regularly sampled and are representative of the Gulf of Alaska.
Core areas are all in bottom depth at or greater than 100 and less than 150
meters. Salintiy or temperature data from the surface to 100m will be shown.
core_stations are as follows: "Line 8 FOX", "Line 8", and "Semidi".
Line 8 FOX is a set of six stations  with a bounding box of `57.52, -155.2,
57.72, -154.85`. Line 8 are the 4 core stations of Line 8 FOX. Semidi are
5 stations centrally located in the Semidi area and have been consistently
sampled, with a bounding box of '55.5, -157.59, 56.5, -157.46'.}

\item{plot_type}{will accept on of two quoted characters "temperature" or
"salinity".}

\item{fastcat_data}{An optional argument if you want to add the current years
fastcat data. Supply the path and .csv file name to the current years
fastcat data. This must be in the format created by the FastrCAT function
make_dataframe_fc.}
}
\value{
A depth by year tile plot of temperature or salinity. The plot will
be written to the folder designated by the historical data file path. The
plot will be in .png format.
}
\description{
Creates either temperature or salinity plots of core EcoFOCI
stations in the Gulf of Alaska. Each plot displays the average temperature
or salinity for each meter of depth of the core stations for each year.
Post 2010, these core stations were only sampled in odd numbered years. In
the future more core areas will be added.
}
