% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts_fc.R
\name{plot_ts_fc}
\alias{plot_ts_fc}
\title{Temperature/Salinity by Depth Plots}
\usage{
plot_ts_fc(current_path)
}
\arguments{
\item{current_path}{The path to directory where dataframe created from
make_dataframe_fc() is located.}
}
\value{
A plot of temperature and salinity by depth for each
station of a cruise. Plots are written in the plot folder and
are in the .png format.
}
\description{
Once make_dataframe_fc has been run, then plot_ts_fc
can be used. This function creates a depth by salinity and temperature
plot for each station. These are all .png files which will be located
in the plot folder within the current folder. If a plot folder hasn't been
created, it will create on. It only needs to be run
once to generate a plot for each station. Each dot is a data point.
Check the profile for each station/haul.
}
