% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fc.R
\name{map_fc}
\alias{map_fc}
\title{Mapping of FastCAT data}
\usage{
map_fc(current_path, map_type = "Station", depth_range = NA)
}
\arguments{
\item{current_path}{The path to the directory where the .csv file
generated by make_dataframe_fc is located.}

\item{map_type}{Determines the type of map which will be returned. The
default map is a station map. Map types are Station: returns map of
station locations, Sample intensity: returns a map of sampling intesity
for each 0.3 decimal degrees which is around 30km for the regions that are
sampled hexagon, Salinity: returns an
map using a least-squares gridding method for the average of the depth_range
specified. If depth_range remains as NA, then the entire water column will
be averaged.Temperature: returns a map using a least-squares gridding method
for the average of the depth_range specified. If depth_range remains as NA,
then the entire water column will be averaged. Map types are as follows and
should be quoted: "Station", "Sample Intensity", "Salinity", and
"Temperature".}

\item{depth_range}{The desired depth range for either the "Salinity" or
"Temperature". If the parameter is left as NA, then the entire water
column will be averages.Takes a two value vector, a minimum and maximum of
desired depth range.}
}
\value{
A station map
}
\description{
Once make_dataframe_fc() has been run, then map_fc can be
used. This function makes a station map from the dataframe generated.
The rendered map will be located in the plot folder within the current
folder.
}
